Switch history primary key to bigserial

+++ spec/support/matchers/column.rb
@@ -87,7 +87,7 @@ module ChronoTest::Matchers
         super([
           ['validity',    'tsrange'],
           ['recorded_at', 'timestamp without time zone'],
-          ['hid',         'integer']
+          ['hid',         'bigint']
         ], history_schema)
       end
     end
